upgrading syntax to Java 8

+++ src/main/java/org/javarosa/xpath/expr/XPathArithExpr.java
@@ -43,8 +43,8 @@ public class XPathArithExpr extends XPathBinaryOpExpr {
     }
 
     public Object eval (DataInstance model, EvaluationContext evalContext) {
-        double aval = XPathFuncExpr.toNumeric(a.eval(model, evalContext)).doubleValue();
-        double bval = XPathFuncExpr.toNumeric(b.eval(model, evalContext)).doubleValue();
+        double aval = XPathFuncExpr.toNumeric(a.eval(model, evalContext));
+        double bval = XPathFuncExpr.toNumeric(b.eval(model, evalContext));
 
         double result = 0;
         switch (op) {
